[CAKER-12] Adding open(URL) capability for macOS
