+++ Package.swift
@@ -15,11 +15,15 @@ let package =
         targets: ["SwiftUIExtensions"]
       ),
     ],
-    dependencies: [],
+    dependencies: [
+      .package(name: "SwiftFoundationExtensions", url: "https://github.com/xiiagency/SwiftFoundationExtensions", .branchItem("main")),
+    ],
     targets: [
       .target(
         name: "SwiftUIExtensions",
-        dependencies: []
+        dependencies: [
+          "SwiftFoundationExtensions",
+        ]
       ),
       // NOTE: Re-enable when tests are added.
 //      .testTarget(
